Add test to prompt service and rename service
