+++ src/AnswersPage.test.jsx
@@ -7,7 +7,7 @@ import { useSelector } from 'react-redux';
 import AnswersPage from './AnswersPage';
 
 jest.mock('react-redux');
-jest.mock('./services/speechRecognition.js');
+jest.mock('./services/speechRecognitionService.js');
 
 describe('AnswersPage', () => {
   const answer = {
