+++ src/services/promptService.test.js
@@ -1,0 +1,7 @@
+import prompts from '../data/prompts';
+
+import { getNextPrompt } from './promptService';
+
+test('getNextPrompt', () => {
+  expect(prompts).toContain(getNextPrompt());
+});
